Added new WIP content data types to the solution in a separate project.
Each content data type is defined in an XML schema (.xsd) which describes the structure and basic data types used in the content data type. Another XML schema defines the structure of the content XML files.

Apart from the XML schemas, each content data type is also represented by a C# class. Each class is implemented as a partial class in two separate files. One of these bears the class name, and contains hand-coded methods and properties not involved in the serialization of the content files. The other file is common to all the content types, and contains the auto-generated serializable C# representation of the XML content formats. This division of the different class parts ensures that new serializable classes can be auto-generated without affecting the manually implemented methods used in the classes.

None of the new content data types are used by the engine yet.