Added a content file editor tool for the new content data types.
The tool is very much a work in progress, and will need to be expanded as new content data types are introduced. I realize I am not a very good GUI designer, so anyone who feels they can improve the user interface are welcome to do so. Still, the tool is usable in its current form, and does make quick work of generating content files. A new animation will typically only take a couple of minutes to make, provided that the needed graphics are available.

Be aware that the logic of the new content data types requires that all graphics are henceforth stored in tile sets.